# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TravelBuddi API",
        "description": "A product of Appetiser",
        "contact": {
            "name": "Iazell",
            "email": "iazell.raper@appetiser.com.au"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{ip or domain}/api/v1",
            "description": "Base Endpoint"
        }
    ],
    "paths": {
...
        "/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get All Items",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiItemsController::index",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Items"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Submit a New Item",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiItemsController::submitNewItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "place_id": {
                                        "description": "Place ID",
                                        "type": "string",
                                        "example": "ChIJKwdvi4-Z4jAR-GnYG7XAGII"
                                    },
                                    "city_id": {
                                        "description": "City ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "item_name": {
                                        "description": "Item Name",
                                        "type": "string",
                                        "example": "Eiffel Tower"
                                    },
                                    "recommended_price": {
                                        "description": "Recommended Price",
                                        "type": "double",
                                        "example": "99.99"
                                    },
                                    "image_url": {
                                        "description": "Item's Image URL",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Items"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/items/user/{user_id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get Followed Items",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiItemsController::getContributedItems",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "User's followed items"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/items/{item_id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get Item",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiItemsController::getItem",
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Items"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}